refactor(connectors): make normalized contracts domain-owned
