+++ src/agent_memory_orchestrator/domain/connectors/__init__.py
@@ -2,22 +2,7 @@
 
 from __future__ import annotations
 
-from .events import ConnectorEvent
-from .events import SlackMessage
-from .events import finalize_connector_event
-from .events import message_to_connector_event
-from .events import parse_message_envelope
-from .events import should_capture_message
-from .events import should_reply_message
+from .models import ConnectorEvent
 from .responses import ConnectorResponse
 
-__all__ = [
-    "ConnectorEvent",
-    "ConnectorResponse",
-    "SlackMessage",
-    "finalize_connector_event",
-    "message_to_connector_event",
-    "parse_message_envelope",
-    "should_capture_message",
-    "should_reply_message",
-]
+__all__ = ["ConnectorEvent", "ConnectorResponse"]
