+++ src/agent_memory_orchestrator/domain/connectors/responses.py
@@ -1,7 +1,16 @@
-"""Connector response domain contracts."""
+"""Source-neutral connector response contracts."""
 
 from __future__ import annotations
 
-from ...extensions.contracts.connector import ConnectorResponse
+from dataclasses import dataclass, field
+from typing import Any, Mapping, Sequence
+
+
+@dataclass(slots=True, frozen=True)
+class ConnectorResponse:
+    ok: bool
+    messages: Sequence[str] = field(default_factory=tuple)
+    diagnostics: Mapping[str, Any] = field(default_factory=dict)
+
 
 __all__ = ["ConnectorResponse"]
