+++ src/agent_memory_orchestrator/integrations/connectors/base.py
@@ -1,29 +1,7 @@
+"""Compatibility facade for the domain-owned normalized connector event."""
+
 from __future__ import annotations
 
-from dataclasses import dataclass, field
-from typing import Any
+from ...domain.connectors.models import ConnectorEvent
 
-
-@dataclass(slots=True, frozen=True)
-class ConnectorEvent:
-    """Normalized event emitted by an external connector before AMO evidence capture."""
-
-    connector: str
-    source_app: str
-    external_id: str
-    session_id: str
-    event_type: str
-    content: str
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-    def as_evidence_payload(self) -> dict[str, Any]:
-        return {
-            "connector": self.connector,
-            "source_app": self.source_app,
-            "external_id": self.external_id,
-            "session_id": self.session_id,
-            "event_type": self.event_type,
-            "message": self.content,
-            "content": self.content,
-            "metadata": self.metadata,
-        }
+__all__ = ["ConnectorEvent"]
